Remove the __iter__ on CTX
As far as I can tell, this isn't necessary. It was originally copied
from the asyncio.Future class when I first made ChildOfFuture like a
decade ago. As far as I can tell, and maybe I'm wrong, that was
necessary on asyncio.Future for compatibility reasons, but isn't
necessary on custom classes that pretend to be futures